slacker/requirements.txt
--index-url=https://artifactory.corp.onapsis.com/artifactory/api/pypi/osp-python/simple
slackclient==2.1.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
psycopg2-binary==2.8.3
celery==4.3.0
redis==3.2.1
pytz==2019.1
requests==2.20.1
dateparser==0.7.1
beautifulsoup4==4.7.1
lxml==4.3.2
loguru==0.2.5
python-dotenv==0.10.2
google-api-python-client==1.7.11
google-auth-oauthlib==0.4.0
google-auth-httplib2==0.0.3
unidecode==1.1.1
pillow==6.1.0
docopt==0.6.2
pyopenssl==19.1.0
cryptography>=2.8
